Faster app start by moving initialization to background tasks

* merge

* merge

* merge

* Move and rename Equipmentchooser

* remove unused equipmentchooser

* Asynchronously load imaging tab layout and show a loading spinner while it is doing so

* remove accidental change

* Fix incorrect framing command initialization

* merge

* Merge branch 'develop' into feature/faststart

+++ NINA/Utility/Astrometry/WorldCoordinateSystem.cs
@@ -140,7 +140,7 @@ namespace NINA.Utility.Astrometry {
         /// <param name="pixelX">x pixel value</param>
         /// <param name="pixelY">y pixel value</param>
         /// <returns>Coordinates at position x|y</returns>
-        public Coordinates GetCoordinates(int pixelX, int pixelY) {
+        public Coordinates GetCoordinates(double pixelX, double pixelY) {
             return Coordinates.Shift(pixelX - Point.X, pixelY - Point.Y, Rotation, PixelScaleX, PixelScaleY, Coordinates.ProjectionType.Gnomonic);
         }
     }
